requirements.txt
kafka-python
psycopg2-binary
python-dotenv
yfinance
pandas
streamlit
plotly
